app/src/androidTest/java/com/quartzo/topratedmovies/provider/MovieContractTest.java
/*
 * MIT License
 *
 * Copyright (c) 2016 Victor Hugo Montes Neves
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package com.quartzo.topratedmovies.provider;

import android.net.Uri;
import android.support.test.runner.AndroidJUnit4;

import org.junit.Test;
import org.junit.runner.RunWith;

import static junit.framework.Assert.assertEquals;

/**
 * Created by victoraldir on 21/11/2016.
 */
@RunWith(AndroidJUnit4.class)
public class MovieContractTest {

    @Test
    public void shouldGetIdFromUI() {

        int id = 100;

        Uri uri = MovieContract.MovieEntry.buildMovieIdUri(id);

        assertEquals("ID is different", id, MovieContract.MovieEntry.getIdFromUri(uri));

    }

    @Test
    public void shouldGetPageFromUI() {

        int page = 3;

        Uri uri = MovieContract.MovieEntry.buildMoviePageUri(page);

        assertEquals("ID is different", page, MovieContract.MovieEntry.getPageFromUri(uri));

    }

}